Move autoyast client to lib directory
