+++ src/lib/autoinstall/clients/autoyast.rb
@@ -1,0 +1,160 @@
+# Copyright (c) [2020] SUSE LLC
+#
+# All Rights Reserved.
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms of version 2 of the GNU General Public License as published
+# by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+# more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with this program; if not, contact SUSE LLC.
+#
+# To contact SUSE LLC about this file by physical or electronic mail, you may
+# find current contact information at www.suse.com.
+
+require "yast"
+
+module Y2Autoinstallation
+  module Clients
+    class Autoyast < Yast::Client
+      def initialize
+        Yast.import "Pkg"
+        Yast.import "UI"
+        textdomain "autoinst"
+        Yast.import "Wizard"
+        Yast.import "Mode"
+        Mode.SetMode("autoinst_config")
+
+        Yast.import "Profile"
+        Yast.import "AutoinstConfig"
+        Yast.import "Y2ModuleConfig"
+        Yast.import "Label"
+        Yast.import "Sequencer"
+        Yast.import "Popup"
+        Yast.import "AddOnProduct"
+        Yast.import "CommandLine"
+        Yast.import "AutoInstall"
+
+        Yast.include self, "autoinstall/dialogs.rb"
+        Yast.include self, "autoinstall/conftree.rb"
+        Yast.include self, "autoinstall/wizards.rb"
+      end
+
+      def main
+        if Builtins.size(Y2ModuleConfig.GroupMap) == 0
+          Yast::Wizard.CreateDialog
+          Yast::Popup.Error(_("Error while reading configuration data."))
+          Yast::Wizard.CloseDialog
+          return :abort
+        end
+
+        Yast::Pkg.CallbackImportGpgKey(
+          fun_ref(
+            Yast::AutoInstall.method(:callbackTrue_boolean_map_integer),
+            "boolean (map <string, any>, integer)"
+          )
+        )
+        Yast::Pkg.CallbackAcceptUnknownGpgKey(
+          fun_ref(
+            Yast::AutoInstall.method(:callbackTrue_boolean_string_string_integer),
+            "boolean (string, string, integer)"
+          )
+        )
+        Yast::Pkg.CallbackAcceptFileWithoutChecksum(
+          fun_ref(
+            Yast::AutoInstall.method(:callbackTrue_boolean_string),
+            "boolean (string)"
+          )
+        )
+        Yast::Pkg.CallbackAcceptUnsignedFile(
+          fun_ref(
+            Yast::AutoInstall.method(:callbackTrue_boolean_string_integer),
+            "boolean (string, integer)"
+          )
+        )
+
+        @cmdline = {
+          "id"         => "autoyast2",
+          "help"       => _("AutoYaST"),
+          "guihandler" => fun_ref(method(:AutoSequence), "any ()"),
+          "actions"    => {
+            "file"   => {
+              "handler" => fun_ref(
+                method(:openFile),
+                "boolean (map <string, string>)"
+              ),
+              "help"    => "file operations"
+            },
+            "module" => {
+              "handler" => fun_ref(
+                method(:runModule),
+                "boolean (map <string, string>)"
+              ),
+              "help"    => "module specific operations"
+            }
+          },
+          "options"    => {
+            "filename" => { "type" => "string", "help" => "filename=XML_PROFILE" },
+            "modname"  => { "type" => "string", "help" => "modname=AYAST_MODULE" }
+          },
+          "mappings"   => { "file" => ["filename"], "module" => ["modname"] }
+        }
+
+        # command line options
+        # Init variables
+        @command = ""
+        @flags = []
+        @options = {}
+        @exit = ""
+        @l = []
+
+        @ret = nil
+        @ret = Yast::CommandLine.Run(@cmdline)
+
+        Yast::AddOnProduct.CleanModeConfigSources
+        :exit
+      end
+
+      def openFile(options)
+        options = deep_copy(options)
+        if !Yast::Profile.ReadXML(Ops.get(options, "filename", ""))
+          Yast::Popup.Error(
+            _(
+              "Error while parsing the control file.\n" \
+                "Check the log files for more details or fix the\n" \
+                "control file and try again.\n"
+            )
+          )
+        end
+        Yast::Popup.ShowFeedback(
+          _("Reading configuration data"),
+          _("This may take a while")
+        )
+        Builtins.foreach(Yast::Profile.ModuleMap) do |p, d|
+          # Set resource name, if not using default value
+          resource = Ops.get_string(d, "X-SuSE-YaST-AutoInstResource", "")
+          resource = p if resource == ""
+          Builtins.y2debug("resource: %1", resource)
+          module_auto = Ops.get_string(d, "X-SuSE-YaST-AutoInstClient", "none")
+          rd = Yast::Y2ModuleConfig.getResourceData(d, resource)
+          Yast::WFM.CallFunction(module_auto, ["Import", rd]) if !rd.nil?
+        end
+        Yast::Popup.ClearFeedback
+        AutoSequence()
+        true
+      end
+
+      def runModule(options)
+        options = deep_copy(options)
+        Yast::AutoinstConfig.runModule = Ops.get(options, "modname", "")
+        AutoSequence()
+        true
+      end
+    end
+  end
+end
